% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{list_functions}
\alias{list_functions}
\title{List exported functions of an R package using ellmer and Ollama}
\usage{
list_functions(pkg_name)
}
\arguments{
\item{pkg_name}{Character. The name of the R package.}
}
\value{
Character vector of exported function names.
}
\description{
List exported functions of an R package using ellmer and Ollama
}
